+++ lib/mainImpl.dart
@@ -16,10 +16,9 @@
 
 import 'dart:async';
 import 'dart:collection';
-import 'dart:ui';
+import 'package:web/web.dart' as web;
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/semantics.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter/foundation.dart';
@@ -207,10 +206,6 @@ void mainImpl({ rokwire.ConfigEnvironment? configEnvironment }) async {
     await WebCacheImageService().init();
 
     runApp(App(initializeError: serviceError));
-
-    if (kIsWeb) {
-      SemanticsBinding.instance.ensureSemantics();
-    }
   }, FirebaseCrashlytics().handleZoneError);
 }
 
@@ -298,6 +293,9 @@ class _AppState extends State<App> with NotificationsListener, TickerProviderSta
       NativeCommunicator().dismissLaunchScreen().then((_) {
         _presentLaunchPopup();
       });
+      if (kIsWeb) {
+        _removeEnableAccessibilityButton();
+      }
     });
 
     super.initState();
@@ -596,6 +594,21 @@ class _AppState extends State<App> with NotificationsListener, TickerProviderSta
       setState(() {
         _contentAlert = contentAlert;
       });
+    }
+  }
+
+  void _removeEnableAccessibilityButton() {
+    final nodeList = web.document.querySelectorAll(
+      'flt-semantics-placeholder[aria-label="Enable accessibility"]',
+    );
+    int nodesCount = nodeList.length;
+    if (nodesCount > 0) {
+      for (var i = 0; i < nodesCount; i++) {
+        final node = nodeList.item(i);
+        if (node != null && node.parentNode != null) {
+          node.parentNode!.removeChild(node);
+        }
+      }
     }
   }
 }